Add errors messages

+++ shop_qraphql/mutations/objects_mutations/bouquet_mutations.py
@@ -1,6 +1,7 @@
 import graphene.types
 from ...objects.objects_types import *
 from ..mutation_pay_load import MutationPayLoad
+from ...form_checks import is_link
 
 
 class BouquetInput(graphene.InputObjectType):
@@ -8,6 +9,24 @@ class BouquetInput(graphene.InputObjectType):
     price = graphene.Float()
     link_on_photo = graphene.String()
     seller_id = graphene.Int()
+
+
+def check_link_on_photo(input, id):
+    if not is_link(input.link_on_photo):
+        return "Field 'link_on_photo' expected for link, but received '{}'.".format(input.link_on_photo)
+    return None
+
+
+def check_id(input, id):
+    if Bouquet.objects.get(pk=id) is None:
+        return "Bouquet with id '{}' is not exist.".format(id)
+    return None
+
+
+def check_seller_id(input, id):
+    if Seller.objects.get(pk=input.seller_id) is None:
+        return "Seller with id '{}' is not exist.".format(id)
+    return None
 
 
 class CreateBouquet(MutationPayLoad, graphene.Mutation):
@@ -19,9 +38,13 @@ class CreateBouquet(MutationPayLoad, graphene.Mutation):
 
     @staticmethod
     def mutate(root, info, input):
-        bouquet = Bouquet(**input)
-        bouquet.save()
-        return CreateBouquet(bouquet=bouquet, bouquet_id=bouquet.id)
+        checkers = [check_seller_id, check_link_on_photo]
+        errors = MutationPayLoad.check_errors(checkers, input=input)
+        if not errors:
+            bouquet = Bouquet(**input)
+            bouquet.save()
+            return CreateBouquet(bouquet=bouquet, bouquet_id=bouquet.id, errors=errors)
+        return CreateBouquet(errors=errors)
 
 
 class UpdateBouquet(MutationPayLoad, graphene.Mutation):
@@ -34,13 +57,17 @@ class UpdateBouquet(MutationPayLoad, graphene.Mutation):
 
     @staticmethod
     def mutate(root, info, id, input):
-        bouquet = Bouquet.objects.get(pk=id)
-        bouquet.name = input.name
-        bouquet.seller = Seller.objects.get(pk=input.seller_id)
-        bouquet.price = input.price
-        bouquet.link_on_photo = input.link_on_photo
-        bouquet.save()
-        return UpdateBouquet(bouquet=bouquet, bouquet_id=bouquet.id)
+        checkers = [check_seller_id, check_link_on_photo, check_id]
+        errors = MutationPayLoad.check_errors(checkers, input=input, id=id)
+        if not errors:
+            bouquet = Bouquet.objects.get(pk=id)
+            bouquet.name = input.name
+            bouquet.seller = Seller.objects.get(pk=input.seller_id)
+            bouquet.price = input.price
+            bouquet.link_on_photo = input.link_on_photo
+            bouquet.save()
+            return CreateBouquet(bouquet=bouquet, bouquet_id=bouquet.id, errors=errors)
+        return UpdateBouquet(errors=errors)
 
 
 class DeleteBouquet(MutationPayLoad, graphene.Mutation):
@@ -49,6 +76,9 @@ class DeleteBouquet(MutationPayLoad, graphene.Mutation):
 
     @staticmethod
     def mutate(root, info, id):
-        bouquet = Bouquet.objects.get(pk=id)
-        bouquet.delete()
-        return DeleteBouquet()
+        checkers = [check_id]
+        errors = MutationPayLoad.check_errors(checkers, id=id)
+        if not errors:
+            bouquet = Bouquet.objects.get(pk=id)
+            bouquet.delete()
+        return DeleteBouquet(errors=errors)
