+++ shop_qraphql/mutations/objects_mutations/customer_mutations.py
@@ -2,11 +2,24 @@ import graphene
 
 from ...objects.objects_types import *
 from ..mutation_pay_load import MutationPayLoad
+from ...form_checks import is_mail
 
 
 class CustomerInput(graphene.InputObjectType):
     name = graphene.String()
     mail = graphene.String()
+
+
+def check_mail(input, id):
+    if not is_mail(input.mail):
+        return "Field 'mail' expected for mail address, but '{}' is not valid mail address.".format(input.mail)
+    return None
+
+
+def check_id(input, id):
+    if not Customer.objects.filter(pk=id):
+        return "Customer with id '{}' is not exist.".format(id)
+    return None
 
 
 class CreateCustomer(MutationPayLoad, graphene.Mutation):
@@ -18,9 +31,13 @@ class CreateCustomer(MutationPayLoad, graphene.Mutation):
 
     @staticmethod
     def mutate(root, info, input):
-        customer = Customer(**input)
-        customer.save()
-        return CreateCustomer(customer=customer, customer_id=customer.id)
+        checkers = [check_mail]
+        errors = MutationPayLoad.check_errors(checkers, input=input)
+        if not errors:
+            customer = Customer(**input)
+            customer.save()
+            return CreateCustomer(customer=customer, customer_id=customer.id, errors=errors)
+        return CreateCustomer(errors=errors)
 
 
 class UpdateCustomer(MutationPayLoad, graphene.Mutation):
@@ -33,11 +50,16 @@ class UpdateCustomer(MutationPayLoad, graphene.Mutation):
 
     @staticmethod
     def mutate(root, info, id, input):
-        customer = Customer.objects.get(pk=id)
-        customer.name = input.name
-        customer.mail = input.mail
-        return UpdateCustomer(customer=customer, customer_id=customer.id)
-    
+        checkers = [check_mail, check_id]
+        errors = MutationPayLoad.check_errors(checkers, input=input, id=id)
+        if not errors:
+            customer = Customer.objects.get(pk=id)
+            customer.name = input.name
+            customer.mail = input.mail
+            customer.save()
+            return UpdateCustomer(customer=customer, customer_id=customer.id)
+        return UpdateCustomer(errors=errors)
+
     
 class DeleteCustomer(MutationPayLoad, graphene.Mutation):
     class Arguments:
@@ -45,6 +67,9 @@ class DeleteCustomer(MutationPayLoad, graphene.Mutation):
 
     @staticmethod
     def mutate(root, info, id):
-        customer = Customer.objects.get(pk=id)
-        customer.delete()
-        return DeleteCustomer()
+        checkers = [check_id]
+        errors = MutationPayLoad.check_errors(checkers, id=id)
+        if not errors:
+            customer = Customer.objects.get(pk=id)
+            customer.delete()
+        return DeleteCustomer(errors=errors)
